finance calculator modified errors

+++ python_introduction/finance_calculator.py
@@ -2,5 +2,5 @@
 monthly_income = int(input("Enter your monthly income: "))
 monthly_expenses = int(input("Enter your monthly Expenses: "))
 monthly_saving = monthly_income - monthly_expenses
-projected_saving = monthly_saving * 12 + (monthly_saving * 12 * 0.05)
+projected_saving = float(monthly_saving * 12 + (monthly_saving * 12 * 0.05))
 print("Your projected saving is:", projected_saving)
